+++ addOns/customactivescan/src/main/java/org/zaproxy/zap/extension/customactivescan/model/PauseActionObject.java
@@ -10,14 +10,10 @@ public class PauseActionObject {
 
     private int pauseCounter = 0;
     private boolean isTerminated = false;
-    private long prevNanoTime = -1;
-    private Random random;
 
     public PauseActionObject() {
         pauseCounter = 0;
         isTerminated = false;
-        prevNanoTime = -1;
-        random = new Random();
     }
 
     synchronized void onceWaiter() {
@@ -143,25 +139,4 @@ public class PauseActionObject {
         }
         return stateString;
     }
-
-    public void waitUntilSpecifiedTimePassed(CustomScanJSONData.ScanRule selectedScanRule) {
-        long mSecWaitTime = selectedScanRule.getIdleTime(this.random);
-        long currentNanoTime = System.nanoTime();
-        if (this.prevNanoTime != -1) {
-            long nanoWaitTime = mSecWaitTime * 1000000;
-            long lapseNanoTime = currentNanoTime - this.prevNanoTime;
-            if (lapseNanoTime < nanoWaitTime) {
-                // wait until nanoWaitTime passes
-                long sleepMSecTime = Math.round((double)(nanoWaitTime - lapseNanoTime) / 1000000);
-                try {
-                    Thread.sleep(sleepMSecTime);
-                } catch (Exception ex) {
-
-                }
-            }
-        }
-        this.prevNanoTime = System.nanoTime();
-    }
-
-
 }
